% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Amatrix}
\alias{Amatrix}
\title{Amatrix}
\usage{
Amatrix(X1, X2)
}
\arguments{
\item{X1}{model matrix for the assumed model}

\item{X2}{model matrix for the additional terms in the true model}
}
\value{
The alias matrix
}
\description{
Calculates an alias matrix for a fractional factorial experiment
}
\examples{
library(FrF2)
design <- FrF2(8, 5, factor.names = paste0("x", 1:5))
X <- model.matrix(~(.) ^ 2, data = design)
X1 <- X[, 1:6]
X2 <- X[, 7:16]
round(Amatrix(X1, X2), 2)
}
