% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{borehole}
\alias{borehole}
\title{borehole}
\usage{
borehole(xx)
}
\arguments{
\item{x}{a vector of length eight giving the values for each of the inputs.
See the reference url for the ranges of each input}
}
\description{
Function that simulates the response from mathematical model of water flow rate
through a borehole
}
\examples{
x <- c(0.12650025, 897.8873, 97266.21, 1041.3891, 76.17152, 745.6965, 1608.609, 11853.055)
borehole(x)
}
\references{
\url{http://www.sfu.ca/~ssurjano/}
}
