% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cirfab}
\alias{cirfab}
\title{cirfab}
\format{A data frame with 16 rows and 5 variables:
\describe{
  \item{\code{x1}}{rotation method}
  \item{\code{x2}}{nozzle position}
  \item{\code{x3}}{deposition temperature}
  \item{\code{x4}}{deposition time}
  \item{\code{ybar}}{thickness of the epitaxial layer (micrometres) - average from six wafers}
}}
\source{
Wu, C.F.J. and Hamada, M.S. (2009). Experiments: Planning, Analysis, and Optimization. 2nd ed. Wiley, New York. Page 155
}
\usage{
cirfab
}
\description{
\eqn{2^4} full factorial design and response from an experiment fabricating integrated circuit boards.
Each factor has coded levels \code{-1,+1}
}
\keyword{datasets}
