% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contr.twolevel}
\alias{contr.twolevel}
\title{contr.twolevel}
\usage{
contr.twolevel()
}
\description{
Contrasts for two-level factors with levels -1 and +1 (recoding factor levels)
}
\examples{
x <- factor(letters[1:2])
contrasts(x) <- contr.twolevel()
}
