% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebm}
\alias{ebm}
\title{ebm}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{\code{y}}{mean temperature}
  \item{\code{x1}}{solar constant}
  \item{\code{x2}}{non-ice albedo}
}}
\source{
The response variable was obtained by running the Energy Balance Model available from \url{https://wiki.aston.ac.uk/foswiki/bin/view/MUCM/SurfebmModel}
}
\usage{
ebm
}
\description{
A maximin Latin hypercube design with two variables and one response
}
\keyword{datasets}
