% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flighttime}
\alias{flighttime}
\title{flighttime}
\usage{
flighttime(w, r, t, d, theta0 = exp(0.102), theta1 = 1.9, phi = 0.1,
  int = 0)
}
\arguments{
\item{w}{rotor width [0.03, 0.09]m}

\item{r}{rotor length [0.07, 0.12]m}

\item{t}{tail length [0.07, 0.12]m}

\item{d}{paper density [0.06, 0.12] kgm^-2}

\item{phi}{standard deviation of response distribution}
}
\value{
Simulated helicopter flight time (seconds)
}
\description{
Function to simulate the response (flight time) from the paper helicopter experiment, using a semi-physical model based on dimensional analysis
}
\details{
The flight time is simulated from a normal distribution with mean derived from a dimensional analysis regression model and standard deviation phi.

Other parameters (theta0, theta1 and int) are regression parameters.
}
\examples{
x <- c(0.03, 0.12, 0.12, 0.06)
flighttime(w = x[, 1], r = x[, 2], t = x[, 3], d = x[, 4])
}
\references{
\url{https://www.paperhelicopterexperiment.com}

Box, G. E. P. and P. Y. T. Liu (1999). Statistics as a catalyst to learning by scientific method part I - an example. Journal of Quality Technology 31, 1–15.

Shen, W., T. Davis, D. K. J. Lin, and C. J. Nachtsheim (2014). Dimensional analysis and its applications in statistics. Journal of Quality Technology 46, 185–198.

Woods, D. C., Overstall, A. M., Adamou, M. and Waite, T. W. (2017). Bayesian design of experiments for generalised linear models and dimensional analysis with industrial and scientific application (with discussion). Quality Engineering, 29, 91-118.
}
