% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.minusplus}
\alias{is.minusplus}
\title{is.minusplus}
\usage{
is.minusplus(x)
}
\arguments{
\item{x}{a number or vector}
}
\description{
Simple utlity function to check if a number is in the set \{-1, +1\}
or if every element of a vector is in this set
}
