% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{materials.response}
\alias{materials.response}
\title{materials.response}
\usage{
materials.response(x)
}
\arguments{
\item{x}{a vector of length six giving the values in \{-1,+1\} for each of the factors}
}
\description{
Function that simulates the response from a materials wear testing experiment
}
\examples{
x <- c(-1, +1, -1, +1, +1, -1)
materials.response(x)
}
