% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssd}
\alias{ssd}
\title{ssd}
\format{A data frame with 10 rows and 16 variables:}
\usage{
ssd
}
\description{
A Bayesian D-optimal supersaturated design with 16 factors and 10 runs
Each factor has coded levels \code{-1,+1}
}
\keyword{datasets}
